+++ src/applyPatches.ts
@@ -56,18 +56,28 @@ export default function findPatchFiles(appPath: string) {
 
 export function applyPatch(patchFilePath: string) {
   try {
-    spawnSafeSync("git", ["apply", "--check", patchFilePath], {
+    spawnSafeSync(
+      "git",
+      ["apply", "--check", "--unsafe-paths", patchFilePath],
+      {
+        logStdErrOnError: false,
+      },
+    )
+    spawnSafeSync("git", ["apply", "--unsafe-paths", patchFilePath], {
       logStdErrOnError: false,
     })
-    spawnSafeSync("git", ["apply", patchFilePath], { logStdErrOnError: false })
   } catch (e) {
     // patch cli tool has no way to fail gracefully if patch was already
     // applied, so to check, we need to try a dry-run of applying the patch in
     // reverse, and if that works it means the patch was already applied
     // sucessfully. Otherwise the patch just failed for some reason.
-    spawnSafeSync("git", ["apply", "--reverse", "--check", patchFilePath], {
-      logStdErrOnError: false,
-    })
+    spawnSafeSync(
+      "git",
+      ["apply", "--reverse", "--check", "--unsafe-paths", patchFilePath],
+      {
+        logStdErrOnError: false,
+      },
+    )
   }
 }